+++ resources/views/expenses/_form.blade.php
@@ -59,7 +59,8 @@
     <div class="col-md-6">
         <x-form-select label="Cuenta / medio de pago" name="account_id" :options="$accounts"
                        valueKey="id" labelKey="name" :selected="$expense?->account_id"
-                       placeholder="Selecciona una cuenta" required />
+                       placeholder="Selecciona una cuenta" required
+                       smartSelect="expense_account" />
     </div>
     {{-- 4. Fecha --}}
     <div class="col-md-6">